Revert "change-component-to-page"
